requirements.txt
autogen_agentchat==0.4.7
autogen_core==0.4.7
autogen_ext==0.4.7
diskcache==5.6.3
fastapi==0.115.9
panel==1.6.1
param==2.2.0
pydantic==2.10.6
python-dotenv==1.0.1
uvicorn==0.34.0
